requirements.txt
neat_python==1.1.0
pygame==2.6.1
